requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
SQLAlchemy==2.0.4
Flask-SQLAlchemy==3.0.3
Jinja2==3.1.2
MarkupSafe==2.1.2
itsdangerous==2.1.2
click==8.1.3
gunicorn==20.1.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
